Set color of status bar and navigation bar.
This patchset sets the color of the status bar to the same color
as the toolbar background. The text and icon color is adapted
to be black or white depending on dark or light theme.

For devices from API 27 and higher, also the color of the
navigation bar is set. For older APIs, this is not done
since we cannot overwrite the color of the symbols such that
they are still visible for light colors.