Edit-instrument has now a custom long click
There were issues with long-clicks of the edit-icon in the
instruments view. By reducing the long-click time, the long
click occurs before the underlying RecyclerView can "steal"
it. Still to be tested, if this really fixes the issue.

Additionally, the InstrumentSectionAdapter is moved to the
instruments-folder since this is where is belongs to.